+++ src/web/public/js/settings.js
@@ -1,6 +1,7 @@
 import { api } from './api.js';
 import { showToast, escapeHtml } from './utils.js';
 import * as Notifications from './notifications.js';
+import * as Fonts from './fonts.js';
 import { t as i18nT, tf as i18nTf } from './i18n.js';
 import { openSheet, confirmSheet, promptSheet } from './sheet.js';